requirements.txt
python-smarttub>=0.0.35
aiohttp>=3.9
